feat(comments): authing with user id

+++ app/Controllers/Http/CommentsController.ts
@@ -1,33 +1,43 @@
 import { prisma } from '@ioc:Adonis/Addons/Prisma'
 import type { HttpContextContract } from '@ioc:Adonis/Core/HttpContext'
 import { Comment } from 'App/Models/Comment'
+import User, { UserDto } from 'App/Models/User'
 import { QueryItems } from 'App/utils/QueryItems'
-
-type CommentQuery = {
-  recipeId: string
-}
 
 export default class CommentsController {
   public async index({ request }: HttpContextContract): Promise<Comment[]> {
-    const queryItems: CommentQuery = QueryItems.parse(request.parsedUrl.query ?? '')
-    const { recipeId } = queryItems
-    if (!recipeId || recipeId.length === 0) throw new Error('missing recipeId')
+    const recipeId = QueryItems.parse(request.parsedUrl.query ?? '')['recipeId']
+    if (!recipeId) throw new Error('missing recipeId')
     const recipe = await prisma.comment.findMany({
       where: { recipeId },
-      select: { text: true, createdAt: true, user: { select: { id: true, name: true } } },
+      select: { text: true, createdAt: true, userId: true },
     })
-    return recipe.map(({ text, createdAt, user }) => {
+    const userIdList = recipe.map((r) => r.userId)
+    const users = await User.query().whereIn('idmain', userIdList).select(['idmain', 'name'])
+    return recipe.map(({ text, createdAt, userId }) => {
       return {
         text,
         createdAt: createdAt.toISOString(),
-        user,
+        user: this.findUser(users, userId),
       }
     })
   }
 
+  private findUser(users: User[], userId: string): UserDto {
+    const user = users.find(({ idmain }) => idmain.localeCompare(userId) === 0)
+    if (!user) return {}
+    const { idmain, name } = user
+    return {
+      id: idmain,
+      name,
+    }
+  }
+
   public async store({ request, response }: HttpContextContract): Promise<{ id: string }> {
     const input: Record<string, string> = request.body()
-    const { recipeId, userId, text }: Comment = input
+    const userId = QueryItems.parse(request.parsedUrl.query ?? '')['userId']
+    if (!userId) throw new Error('missing user id')
+    const { recipeId, text }: Comment = input
     const comment = await prisma.comment.create({ data: { recipeId, userId, text } })
     response.status(201)
     return {
@@ -35,9 +45,11 @@ export default class CommentsController {
     }
   }
 
-  public async destroy({ params }: HttpContextContract): Promise<{ id: string }> {
+  public async destroy({ request, params }: HttpContextContract): Promise<{ id: string }> {
     const id: string = params.id
-    await prisma.comment.delete({ where: { id } })
+    const userId = QueryItems.parse(request.parsedUrl.query ?? '')['userId']
+    if (!userId) throw new Error('missing user id')
+    await prisma.comment.delete({ where: { id, userId } })
     return {
       id,
     }
